+++ include/vrv/expansionmap.h
@@ -17,6 +17,9 @@
 
 namespace vrv {
 
+class Score;
+class Section;
+
 class ExpansionMap {
 
 public:
@@ -52,6 +55,25 @@ public:
      */
     void ToJson(std::string &output);
 
+    /**
+     * Generate an expan for the score analysing the repeats and endings
+     */
+    void GenerateExpansionFor(Score *score);
+
+    //----------------//
+    // Static methods //
+    //----------------//
+
+    /**
+     * @name Methods to check if a measure yields a repeat start or end
+     */
+    ///@{
+    static bool IsRepeatStart(Measure *measure);
+    static bool IsRepeatEnd(Measure *measure);
+    static bool IsNextRepeatStart(Measure *measure);
+    static bool IsPreviousRepeatEnd(Measure *measure);
+    ///@}
+
 private:
     bool UpdateIDs(Object *object);
 
@@ -61,6 +83,9 @@ private:
 
     /** Ads an id string to an original/notated id */
     bool AddExpandedIDToExpansionMap(const std::string &origXmlId, std::string newXmlId);
+
+    std::string CreateSection(
+        Section *section, const ListOfObjects::iterator &first, const ListOfObjects::iterator &last);
 
 public:
     /** The expansion map indicates which xmlId has been repeated (expanded) elsewhere */