+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## [unreleased]
 * Support for ossia staves
-* Improved MusicXML importer (articulations, spacing, clef changes, tempo spacing)
+* Support for expansion generation in MusicXML importer
+* Improved MusicXML importer (articulations, spacing, clef changes, tempo spacing, coda, segno)
 * Improved `note@head.fill` support
 * Improved SVG styling structure
 * Option `--spacing-ossia` for adjusting the spacing of ossia staves
 * Option `--ossia-hidden` for hiding rendering of ossia staves
+* Option `--expand-first` to expand first `expansion` in the score
+* Option `--expand-generate` to generate `expansion` from measure repeats
 
 ## [5.7.0] – 2025-12-02
 * Support for `@ho` and `@vo` on relevant layer elements and control events